Added version and moved to CLI format
