+++ .goreleaser.yml
@@ -1,10 +1,12 @@
 # ref. https://goreleaser.com/customization/build/
 builds:
   - id: avalanche-network-runner
-    main: ./cmd/avalanche-network-runner
+    main: ./main.go
     binary: avalanche-network-runner
     flags:
       - -v
+    ldflags:
+      - -X 'github.com/ava-labs/avalanche-network-runner/cmd.Version={{.Version}}'
     ignore:
       - goos: darwin
         goarch: 386
