+++ cmd/ping/ping.go
