+++ cmd/root.go
@@ -1,22 +1,25 @@
 // Copyright (C) 2019-2022, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package main
+package cmd
 
 import (
 	"fmt"
 	"os"
 
-	"github.com/ava-labs/avalanche-network-runner/cmd/avalanche-network-runner/control"
-	"github.com/ava-labs/avalanche-network-runner/cmd/avalanche-network-runner/ping"
-	"github.com/ava-labs/avalanche-network-runner/cmd/avalanche-network-runner/server"
+	"github.com/ava-labs/avalanche-network-runner/cmd/control"
+	"github.com/ava-labs/avalanche-network-runner/cmd/ping"
+	"github.com/ava-labs/avalanche-network-runner/cmd/server"
 	"github.com/spf13/cobra"
 )
+
+var Version = ""
 
 var rootCmd = &cobra.Command{
 	Use:        "avalanche-network-runner",
 	Short:      "avalanche-network-runner commands",
 	SuggestFor: []string{"network-runner"},
+	Version:    Version,
 }
 
 func init() {
@@ -31,7 +34,7 @@ func init() {
 	)
 }
 
-func main() {
+func Execute() {
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Fprintf(os.Stderr, "avalanche-network-runner failed %v\n", err)
 		os.Exit(1)
